+++ R/effect_draw.R
@@ -166,7 +166,7 @@ effect_draw <- function(object,
 #'            a comparison will be plotted and  names of objects
 #'            will be used as a plot legend (see examples).
 #' @param transform a transformation to apply to the result, should be an R function;
-#'                  (this is commonly used when calling `group_effects` from other
+#'                  (this is commonly used when calling [group_effects] from other
 #'                  plotting or printing functions)
 #' @return A `ggplot` object.
 #' @import bayesplot
@@ -258,8 +258,3 @@ effect_plot <- function(..., transform=NULL) {
     {if(single_model_flag) theme(legend.position = "none")} +
     {if(n_parameters > 1) facet_wrap(~variable, ncol = 3, scales = "free")}
 }
-
-
-#' @rdname effect_draw
-#' @export
-posterior_predict.baggr <- effect_draw